src/test/java/com/example/controllerTest/RelationshipControllerTest.java
package com.example.controllerTest;

import com.example.controller.RelationshipController;
import com.example.dto.RelationshipDTO;
import com.example.model.Relationship;
import com.example.model.RelationshipPK;
import com.example.model.User;
import com.example.service.RelationshipService;
import com.example.service.UserService;
import com.example.model.request.RequestFriends;
import com.example.model.request.RequestFriendsList;
import com.example.model.request.RequestReciveUpdate;
import com.example.model.request.RequestSubcriber;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.hamcrest.Matchers;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.MediaType;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.test.context.junit4.SpringRunner;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.MvcResult;
import java.util.ArrayList;
import java.util.Arrays;
import java.util.List;
import java.util.Optional;
import static org.hamcrest.Matchers.hasSize;
import static org.hamcrest.core.Is.is;
import static org.mockito.Mockito.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;

@RunWith(SpringRunner.class)
@WebMvcTest(RelationshipController.class)
@AutoConfigureMockMvc(addFilters = false)
@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
public class RelationshipControllerTest {
    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private UserService userService;

    @MockBean
    private RelationshipService relationshipService;

    public static String asJsonString(final Object obj) {
        try {
            return new ObjectMapper().writeValueAsString(obj);
        } catch (Exception e) {
            throw new RuntimeException(e);
        }
    }
    @WithMockUser(username = "len", password = "abc", roles = {"ADMIN"})
    @Test
    public void testGetAllRelationships() throws Exception {
        RelationshipPK relationshipPK = new RelationshipPK("jason@gmail.com", "kati@gmail.com");
        RelationshipDTO relationshipDTO = new RelationshipDTO(relationshipPK, true, false, false);

        RelationshipPK relationshipPK1 = new RelationshipPK("juvia@gmail.com", "david@gmail.com");
        RelationshipDTO relationshipDTO1 = new RelationshipDTO(relationshipPK1, true, false, false);

        List<RelationshipDTO> relationshipDTOList = Arrays.asList(relationshipDTO, relationshipDTO1);
        when(relationshipService.findAllRelationships()).thenReturn(relationshipDTOList);
        MvcResult result = mockMvc.perform(get("/api/relationship"))
                .andExpect(status().isOk())
                .andExpect(content().contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$[0].relationshipPK.userEmail", is("jason@gmail.com")))
                .andExpect(jsonPath("$[1].relationshipPK.userEmail", is("juvia@gmail.com")))
                .andExpect(content().contentType("application/json"))
                .andExpect(jsonPath("$", hasSize(2))).andReturn();
        String content = result.getResponse().getContentAsString();
        Assert.assertEquals("[{\"relationshipPK\":{\"userEmail\":\"jason@gmail.com\",\"friendEmail\":\"kati@gmail.com\"},\"areFriends\":true,\"isSubscriber\":false,\"isBlock\":false},{\"relationshipPK\":{\"userEmail\":\"juvia@gmail.com\",\"friendEmail\":\"david@gmail.com\"},\"areFriends\":true,\"isSubscriber\":false,\"isBlock\":false}]",content);
        verify(relationshipService, times(2)).findAllRelationships();
        verifyNoMoreInteractions(relationshipService);
    }
    @WithMockUser(username = "len", password = "abc", roles = {"ADMIN"})
    @Test
    public void testGetAllRelationshipsEmptyList() throws Exception {
        List<RelationshipDTO> relationshipDTOList = new ArrayList<>();
        when(relationshipService.findAllRelationships()).thenReturn(relationshipDTOList);
        mockMvc.perform(get("/api/relationship"))
                .andExpect(status().isNoContent());
        verify(relationshipService, times(1)).findAllRelationships();
        verifyNoMoreInteractions(relationshipService);
    }
    @Test
    public void testGetRelationship() throws Exception {
        RelationshipPK relationshipPK = new RelationshipPK("jason@gmail.com", "kati@gmail.com");
        RelationshipDTO relationshipDTO = new RelationshipDTO(relationshipPK, true, false, false);
        when(relationshipService.findRelationshipById(relationshipPK)).thenReturn(Optional.of(relationshipDTO));
        MvcResult result = mockMvc.perform(post("/api/relationship/find-relationship-by-id")
                .content(asJsonString(relationshipPK))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.relationshipPK.userEmail", is("jason@gmail.com")))
                .andExpect(content().contentType("application/json")).andReturn();
        String content = result.getResponse().getContentAsString();
        Assert.assertEquals("{\"relationshipPK\":{\"userEmail\":\"jason@gmail.com\",\"friendEmail\":\"kati@gmail.com\"},\"areFriends\":true,\"isSubscriber\":false,\"isBlock\":false}",content);
        verify(relationshipService, times(1)).findRelationshipById(relationshipPK);
        verifyNoMoreInteractions(relationshipService);
    }
    @Test
    public void testGetRelationshipFailByEmailWrongFormat() throws Exception {
        RelationshipPK relationshipPK = new RelationshipPK("jason", "kati");
        mockMvc.perform(post("/api/relationship/find-relationship-by-id")
                .content(asJsonString(relationshipPK))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$.errors[0]", is("error email")))
                .andExpect(jsonPath("$.errors[1]", is("error email")))
                .andExpect(status().isBadRequest())
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testGetRelationshipFailByEmailNull() throws Exception {
        RelationshipPK relationshipPK = new RelationshipPK(null, null);
        mockMvc.perform(post("/api/relationship/find-relationship-by-id")
                .content(asJsonString(relationshipPK))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("not null for email")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testGetRelationshipFailByEmailEmpty() throws Exception {
        RelationshipPK relationshipPK = new RelationshipPK("", "");
        mockMvc.perform(post("/api/relationship/find-relationship-by-id")
                .content(asJsonString(relationshipPK))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("must not be empty")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testBeFriends() throws Exception {
        List<String> listEmail = new ArrayList<>();
        listEmail.add("jason@gmail.com");
        listEmail.add("kati@gmail.com");
        RequestFriends requestFriends = new RequestFriends(listEmail);

        when(relationshipService.beFriends(listEmail.get(0), listEmail.get(1))).thenReturn(true);
        MvcResult result = mockMvc.perform(post("/api/relationship")
                .content(asJsonString(requestFriends))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$.success", is(true)))
                .andExpect(content().contentType("application/json")).andReturn();
        String content = result.getResponse().getContentAsString();
        Assert.assertEquals("{\"success\":true}",content);
        verify(relationshipService, times(1)).beFriends(listEmail.get(0), listEmail.get(1));
        verifyNoMoreInteractions(relationshipService);
    }
    @Test
    public void testBeFriendsFailByEmailWrongFormat() throws Exception {
        List<String> listEmail = new ArrayList<>();
        listEmail.add("jason");
        listEmail.add("kati");
        RequestFriends requestFriends = new RequestFriends(listEmail);

        mockMvc.perform(post("/api/relationship")
                .content(asJsonString(requestFriends))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("email error")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testBeFriendsFailByEmailNull() throws Exception {
        List<String> listEmail = new ArrayList<>();
        listEmail.add(null);
        listEmail.add(null);
        RequestFriends requestFriends = new RequestFriends(listEmail);

        mockMvc.perform(post("/api/relationship")
                .content(asJsonString(requestFriends))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("not null for email")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testBeFriendsFailByEmailEmpty() throws Exception {
        List<String> listEmail = new ArrayList<>();
        listEmail.add("");
        listEmail.add("");
        RequestFriends requestFriends = new RequestFriends(listEmail);

        mockMvc.perform(post("/api/relationship")
                .content(asJsonString(requestFriends))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("must not be empty")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testBeFriendsFailByLackEmail() throws Exception {
        List<String> listEmail = new ArrayList<>();
        listEmail.add("jason@gmail.");
        RequestFriends requestFriends = new RequestFriends(listEmail);
        mockMvc.perform(post("/api/relationship")
                .content(asJsonString(requestFriends))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("size must be 2")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testGetFriendsList() throws Exception {
        User user1 = new User("jason@gmail.com");
        RequestFriendsList requestFriendsList = new RequestFriendsList(user1.getEmail());

        List<String> listEmail = new ArrayList<>();
        listEmail.add("susan@gmail.com");
        listEmail.add("kati@gmail.com");
        when(relationshipService.findFriendsList(requestFriendsList.getEmail())).thenReturn(listEmail);

        MvcResult result = mockMvc.perform(post("/api/relationship/friends-list")
                .content(asJsonString(requestFriendsList))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(jsonPath("$.friends[0]",is("susan@gmail.com")))
                .andExpect(status().isOk())
                .andExpect(content().contentType("application/json")).andReturn();
        String content = result.getResponse().getContentAsString();
        Assert.assertEquals("{\"success\":true,\"friends\":[\"susan@gmail.com\",\"kati@gmail.com\"],\"count\":2}",content);
        verify(relationshipService, times(1)).findFriendsList(requestFriendsList.getEmail());
        verifyNoMoreInteractions(relationshipService);
    }
    @Test
    public void testGetFriendsListFailByEmailWrongFormat() throws Exception {
        User user1 = new User("jason");
        RequestFriendsList requestFriendsList = new RequestFriendsList(user1.getEmail());

        mockMvc.perform(post("/api/relationship/friends-list")
                .content(asJsonString(requestFriendsList))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("email error")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testGetFriendsListFailByEmailNull() throws Exception {
        RequestFriendsList requestFriendsList = new RequestFriendsList(null);

        mockMvc.perform(post("/api/relationship/friends-list")
                .content(asJsonString(requestFriendsList))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("not null for email")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testGetFriendsListFailByEmailEmpty() throws Exception {
        User user1 = new User("");
        RequestFriendsList requestFriendsList = new RequestFriendsList(user1.getEmail());

        mockMvc.perform(post("/api/relationship/friends-list")
                .content(asJsonString(requestFriendsList))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("must not be empty")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testGetCommonFriendsList() throws Exception{
        List<String> listEmail = new ArrayList<>();
        listEmail.add("jason@gmail.com");
        listEmail.add("kati@gmail.com");
        RequestFriends requestFriends = new RequestFriends(listEmail);

        List<String> lstEmail = new ArrayList<>();
        lstEmail.add("david@gmail.com");
        lstEmail.add("susan@gmail.com");

        when(relationshipService.findCommonFriendsList(listEmail)).thenReturn(lstEmail);
        MvcResult result = mockMvc.perform(post("/api/relationship/common-friends-list")
                .content(asJsonString(requestFriends))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(content().contentType("application/json")).andReturn();
        String content = result.getResponse().getContentAsString();
        Assert.assertEquals("{\"success\":true,\"friends\":[\"david@gmail.com\",\"susan@gmail.com\"],\"count\":2}",content);
        verify(relationshipService, times(1)).findCommonFriendsList(listEmail);
        verifyNoMoreInteractions(relationshipService);
    }
    @Test
    public void testGetCommonFriendsListFailByLackEmail() throws Exception{
        List<String> listEmail = new ArrayList<>();
        listEmail.add("jason@gmail.com");
        RequestFriends requestFriends = new RequestFriends(listEmail);

        mockMvc.perform(post("/api/relationship/common-friends-list")
                .content(asJsonString(requestFriends))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("size must be 2")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testGetCommonFriendsListFailByEmailWrongFormat() throws Exception{
        List<String> listEmail = new ArrayList<>();
        listEmail.add("jason");
        listEmail.add("kati");
        RequestFriends requestFriends = new RequestFriends(listEmail);

        mockMvc.perform(post("/api/relationship/common-friends-list")
                .content(asJsonString(requestFriends))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("email error")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testGetCommonFriendsListFailByEmailNull() throws Exception{
        List<String> listEmail = new ArrayList<>();
        listEmail.add(null);
        listEmail.add(null);
        RequestFriends requestFriends = new RequestFriends(listEmail);

        mockMvc.perform(post("/api/relationship/common-friends-list")
                .content(asJsonString(requestFriends))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("not null for email")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testGetCommonFriendsListFailByEmailEmpty() throws Exception{
        List<String> listEmail = new ArrayList<>();
        listEmail.add("");
        listEmail.add("");
        RequestFriends requestFriends = new RequestFriends(listEmail);

        mockMvc.perform(post("/api/relationship/common-friends-list")
                .content(asJsonString(requestFriends))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("must not be empty")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testBeSubscriber() throws  Exception{
        Relationship relationship =new Relationship();

        User user1 = new User("jason@gmail.com");
        User user2 = new User("kati@gmail.com");

        RequestSubcriber requestSubcriber = new RequestSubcriber(user1.getEmail(), user2.getEmail());

        when(relationshipService.beSubscriber(user1.getEmail(), user2.getEmail())).thenReturn(relationship);

        MvcResult result = mockMvc.perform(post("/api/relationship/be-subscriber")
                .content(asJsonString(requestSubcriber))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(content().contentType("application/json")).andReturn();
        String content = result.getResponse().getContentAsString();
        Assert.assertEquals("{\"success\":true}",content);
        verify(relationshipService, times(1)).beSubscriber(user1.getEmail(), user2.getEmail());
        verifyNoMoreInteractions(relationshipService);
    }

    @Test
    public void testBeSubscriberFailByEmailWrongFormat() throws  Exception{
        User user1 = new User("jason");
        User user2 = new User("kati");

        RequestSubcriber requestSubcriber = new RequestSubcriber(user1.getEmail(), user2.getEmail());

        mockMvc.perform(post("/api/relationship/be-subscriber")
                .content(asJsonString(requestSubcriber))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("email error")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testBeSubscriberFailByEmailNull() throws  Exception{
        RequestSubcriber requestSubcriber = new RequestSubcriber(null, null);

        mockMvc.perform(post("/api/relationship/be-subscriber")
                .content(asJsonString(requestSubcriber))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("not null for email")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testBeSubscriberFailByEmailEmpty() throws  Exception{
        RequestSubcriber requestSubcriber = new RequestSubcriber("", "");

        mockMvc.perform(post("/api/relationship/be-subscriber")
                .content(asJsonString(requestSubcriber))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("must not be empty")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testToBlock() throws Exception{
        Relationship relationship =new Relationship();
        User user1 = new User("jason@gmail.com");
        User user2 = new User("kati@gmail.com");
        RequestSubcriber requestSubcriber = new RequestSubcriber(user1.getEmail(), user2.getEmail());

        when(relationshipService.toBlock(user1.getEmail(), user2.getEmail())).thenReturn(relationship);
        MvcResult result = mockMvc.perform(post("/api/relationship/to-block")
                .content(asJsonString(requestSubcriber))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(content().contentType("application/json")).andReturn();
        String content = result.getResponse().getContentAsString();
        Assert.assertEquals("{\"success\":true}",content);
        verify(relationshipService, times(1)).toBlock(user1.getEmail(), user2.getEmail());
        verifyNoMoreInteractions(relationshipService);
    }
    @Test
    public void testToBlockFailByEmailWrongFormat() throws Exception{
        User user1 = new User("jason");
        User user2 = new User("kati");
        RequestSubcriber requestSubcriber = new RequestSubcriber(user1.getEmail(), user2.getEmail());

        mockMvc.perform(post("/api/relationship/to-block")
                .content(asJsonString(requestSubcriber))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("email error")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testToBlockFailByEmailNull() throws Exception{
        RequestSubcriber requestSubcriber = new RequestSubcriber(null, null);

        mockMvc.perform(post("/api/relationship/to-block")
                .content(asJsonString(requestSubcriber))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("not null for email")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testToBlockFailByEmailEmpty() throws Exception{
        User user1 = new User("");
        User user2 = new User("");
        RequestSubcriber requestSubcriber = new RequestSubcriber(user1.getEmail(), user2.getEmail());

        mockMvc.perform(post("/api/relationship/to-block")
                .content(asJsonString(requestSubcriber))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("must not be empty")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testReceiveUpdate() throws Exception{
        User user = new User("jason@gmail.com");
        String text = "this is a text";
        List<String> listReceiveUpload = new ArrayList<>();
        listReceiveUpload.add("kati@gmail.com");
        RequestReciveUpdate requestReciveUpdate = new RequestReciveUpdate(user.getEmail(),text);

        when(relationshipService.findReceiveUpdateList(user.getEmail(),text)).thenReturn(listReceiveUpload);
        MvcResult result = mockMvc.perform(post("/api/relationship/receive-update")
                .content(asJsonString(requestReciveUpdate))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(content().contentType("application/json"))
                .andExpect(jsonPath("$.success", is(true)))
                .andExpect(jsonPath("$.recipients[0]",is("kati@gmail.com")))
                .andExpect(jsonPath("$.*", Matchers.hasSize(2))).andReturn();
        String content = result.getResponse().getContentAsString();
        Assert.assertEquals("{\"success\":true,\"recipients\":[\"kati@gmail.com\"]}",content);
        verify(relationshipService, times(1)).findReceiveUpdateList(user.getEmail(),text);
        verifyNoMoreInteractions(relationshipService);
    }
    @Test
    public void testReceiveUpdateFailByEmailWrongFormat() throws Exception{
        User user = new User("jason");
        String text = "this is a text";
        RequestReciveUpdate requestReciveUpdate = new RequestReciveUpdate(user.getEmail(),text);

        mockMvc.perform(post("/api/relationship/receive-update")
                .content(asJsonString(requestReciveUpdate))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("email error")))
                .andExpect(content().contentType("application/json"));
    }
    @Test
    public void testReceiveUpdateFailByEmailNull() throws Exception{
        String text = "this is a text";
        RequestReciveUpdate requestReciveUpdate = new RequestReciveUpdate(null,text);
        mockMvc.perform(post("/api/relationship/receive-update")
                .content(asJsonString(requestReciveUpdate))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("not null for email")))
                .andExpect(content().contentType("application/json")).andReturn();
    }
    @Test
    public void testReceiveUpdateFailByEmailEmpty() throws Exception{
        User user = new User("");
        String text = "this is a text";
        RequestReciveUpdate requestReciveUpdate = new RequestReciveUpdate(user.getEmail(),text);
        mockMvc.perform(post("/api/relationship/receive-update")
                .content(asJsonString(requestReciveUpdate))
                .contentType(MediaType.APPLICATION_JSON))
                .andExpect(status().isBadRequest())
                .andExpect(jsonPath("$.errors[0]",is("must not be empty")))
                .andExpect(content().contentType("application/json"));
    }
}
